+++ test/unit-tests/function/string/format.test.js
@@ -2,6 +2,7 @@
 import assert from 'assert'
 
 import math from '../../../../src/defaultInstance.js'
+import { format } from '../../../../src/utils/number.js'
 
 describe('format', function () {
   it('should format numbers', function () {
@@ -186,11 +187,15 @@ describe('format', function () {
       assert.strictEqual(math.format(1.25, { notation: 'hex' }), '0x1.4')
       assert.strictEqual(math.bin(-128), '-0b10000000')
       assert.strictEqual(math.bin(-128, 8), '0b10000000i8')
+      assert.strictEqual(math.bin(math.bignumber(-128), math.bignumber(8)), '0b10000000i8')
       assert.strictEqual(math.oct(-128), '-0o200')
       assert.strictEqual(math.oct(-128, 8), '0o200i8')
+      assert.strictEqual(math.oct(math.bignumber(-128), math.bignumber(8)), '0o200i8')
       assert.strictEqual(math.hex(-128), '-0x80')
       assert.strictEqual(math.hex(-128, 8), '0x80i8')
+      assert.strictEqual(math.hex(math.bignumber(-128), math.bignumber(8)), '0x80i8')
     })
+
     it('should throw an error for invalid values', function () {
       assert.throws(function () { math.format(1.25, { notation: 'hex', wordSize: 8 }) }, 'Error: Value must be an integer')
       assert.throws(function () { math.format(1, { notation: 'hex', wordSize: -8 }) }, 'Error: size must be greater than 0')
@@ -228,6 +233,19 @@ describe('format', function () {
       assert.strictEqual(math.format(oneThird, { notation: 'fixed', precision: 0 }), '0')
       assert.strictEqual(math.format(twoThirds, { notation: 'fixed', precision: 0 }), '1')
       assert.strictEqual(math.format(math.bignumber('123456789.123456789'), { notation: 'fixed', precision: 5 }), '123456789.12346')
+    })
+
+    it('should support a Bignumber as precision', function () {
+      assert.strictEqual(math.format(1.2345, math.bignumber(3)), '1.23')
+      assert.strictEqual(math.format(1.2345, { precision: math.bignumber(3) }), '1.23')
+      assert.strictEqual(math.format(math.bignumber('1.2345'), math.bignumber(3)), '1.23')
+      assert.strictEqual(math.format(math.bignumber('1.2345'), { precision: math.bignumber(3) }), '1.23')
+    })
+
+    it('should support a Bignumber as wordSize', function () {
+      const options = { notation: 'hex', wordSize: math.bignumber(32) }
+      assert.strictEqual(math.format(-830, options), '0xfffffcc2i32')
+      assert.strictEqual(math.format(math.bignumber('-830'), options), '0xfffffcc2i32')
     })
 
     describe('engineering notation', function () {
@@ -357,6 +375,20 @@ describe('format', function () {
         assert.strictEqual(math.format(bignumber(777777.77777777), { notation: 'engineering', precision: 1 }), '800e+3')
         assert.strictEqual(math.format(bignumber(777777.77777777), { notation: 'engineering', precision: 2 }), '780e+3')
         assert.strictEqual(math.format(bignumber(-0.000000000777777), { notation: 'engineering', precision: 2 }), '-780e-12')
+      })
+
+      it('should support BigNumber lowerExp and upperExp', function () {
+        const options = {
+          lowerExp: math.bignumber('-2'),
+          upperExp: math.bignumber('2')
+        }
+        assert.strictEqual(format(1, options), '1')
+        assert.strictEqual(format(1e-1, options), '0.1')
+        assert.strictEqual(format(1e-2, options), '0.01')
+        assert.strictEqual(format(1e-3, options), '1e-3')
+        assert.strictEqual(format(1e1, options), '10')
+        assert.strictEqual(format(1e2, options), '1e+2')
+        assert.strictEqual(format(1e3, options), '1e+3')
       })
     })